configure: make pcap_activate optional
Allow configure to pass if pcap_activate is not available. This
is the case on Windows with the msys libpcap which could still
be useful for reading pcap files.

Also make better use of AC_CHECK_LIB to avoid the pcap libs from
being specified multiple times by overriding the if-found
action.