flow: fix timestamp in pseudo packet setup
When Suricata was creating a pseudo packet for flow cleaning
purpose, it was using the timestamp of first packet of the pcap
file when running in pcap mode.

As a result, the timestamp of events triggered during this
flow cleaning phase were completely wrong. As they did often occurs
before the start of the flow itself.

This patch fix this by using flow lastts as timestamp of the packet.
If this is not exact, this is at least really close to reality.

In the case of live traffic, this is also correct as TimeGet function
will get the time when the callback is run which can differ a lot
from the time of the event (which is before timestamp of last seen
packet).