devguide: fix image generation for make dist
Only generate images if the .msc files are newer than the .png files.

During a "make distcheck" the images are generated as part of the "dist"
phase of the build so they exist in distribution archive that
"distcheck" is validating, however generate-images fails here.

During "distcheck" the source tree is read-only as it tests an
out-of-tree build.  In the "distcheck" case the images are there as they
were created during "dist", so only regenerating if the source is newer
fixes this case.

It is still somewhat broken for out of tree builds tho, as sphinx-build
doesn't really have the concept of an out of tree build - we can't
generate the images /over/on/this/path and reference them from
documentation if not under the same root. So its still possible to hit
this case in some uncommon build scenarios.

It might just be easier to regenerate the images as needed and check
them in.