detect/alert: fix same sid alert w same packet bug
with the implementation of configurable packet-alert-max, a bug was
found in the way we added alerts to the packet queue. But the new logic
exposed a possible bug case when the same signature generated more than
one alert for the same packet (from different transactions). This would
lead to Suri call memmove with size 0.