devguide: drop use of script to generate msc imgs
Currently, it seems easier to upload the diagram images to git than to
try to make image generation script work with out of the tree builds
and other corner cases.

This means, however, that one must activelly remember to update msc
diagram files, run the script and re-add new png files, if those ever
need to be updated. To raise awareness to that, a watermark was added
to the diagram images.