dpdk: linker do not remove unnecessary libraries from the binary
Flag `--libs-only-L` of pkg-config should output a path where
libdpdk is stored. This usually happens when DPDK is installed
manually (from source) into a custom location. On the other hand,
there is no output when installing DPDK with a package manager.
For that reason, the linker flag `-Wl,-R,` remain incomplete.

The flag `-Wl,-R,` then binds incorrectly with the following output of
pkg-config. The output of `pkg-config --libs libdpdk` contains linker
flag `--as-need` to remove unnecessary libraries during the linking.
It is silently skipped and the binary is linked with all of the DPDK
libraries.