smtp: bring input out of state to avoid complexity
Create a structure to store input and input length for SMTP which only
consists of primary data types only. This removes the need to store
input in SMTPState and opens path to pass around input and input_len
freely without having to pass the entire SMTPState whenever input is needed.

The struct was created to save the state of input as it is to be used
by SMTPGetLine function in a loop with the updated values each time.

Remove input and input_len from SMTPState and update the tests where
the input_len was checked after parsing.