doxygen: enable caller graph for static funcs
We want to be able to share _some_ calls related to the
unittests functions, but not the whole thing. When documentation for
static functions is enabled in doxygen, the caller graphs can become
huge, for all unittests that are involved in the caller graph will show
up. To prevent this, removing UNITTESTS from PREDEFINED setting was
necessary.

With this change, only the RunUnittests function is shown in caller
graphs.

Related to
Task #5274