detect/ip: switch to radix based address handling
Until now the rule address handling was based on a list of positive
ranges approach, where the address portions of rules would be parsed
and turned into these positive ranges. This was not very efficient
for larger ranges, and especially in complex environments where the
address variables are complex this wasn't efficient.

Additionally, the IP only engine had its own logic based on the radix
tree matching and essentially re-parsed the same address strings.

This is the first commit in a redesign that unifies handling of address
parsing and matching around radix trees.

Each rule now has 4 separate radix trees. 2 per direction for both
IPv4 and IPv6. The rule parser populates these and records their
state as positive or negative match. The tree is somewhat optimized
in that it tries to avoid redundancy.

Additionally, the `detect address map` logic has been renamed to
`ipcache` and moved into its own file.