detect/alert: process all alerts then add to queue
Since we have the final alert queue size when we reach
PacketAlertQueueFinalize, we use a fixed size temporary alert queue to
insert all alerts (except those removed by thresholding). After all
matched alerts are in this queue, we sort it, check if drop or pass
action sigs should be applied, and add the alerts that are left to the
Packet Alerts' queue.

Details:
After processing all alerts, we make sure to save the drop action, if
there was an alert queue expansion failure that could lead to the drop sig not
being appended to the queue.

We also check if the same packet triggers PASS and DROP rules: we
compare sids' internal ids before choosing which action to apply.

Task #5180