Use uniform type for tenant_id
PROBLEM:
The tenant_id property is defined as an int in some structures and as uint32_t in some structures.
In some edge case values of tenant_id (For eg. 2^32), this can cause problems.

SOLUTION:
Use uint32_t for tenant_id through out the code base

This PR is for the redmine  ticket: https://redmine.openinfosecfoundation.org/issues/5186