detect: option to run flowbits on stream
When a flowbit is set on application layer tests, this has
for consequence that there is no point in checking it per packet.
This patch adds the `data` option to the flowbits keyword so
the evaluation is done at the stream level and not at the packet
one.

As a result, using the data option is preventing the bug
https://redmine.openinfosecfoundation.org/issues/2836
to appear.

Ticket: #2836