Fix dns in tcp flow not detected
This fixes issue 4759. The problem is, that when tcp rules are active
in case of a tcp dns connection first the server to client is inspected
and nothing found.
Afterwards the full flow is marked as processed and therefore the dns
query in TCP is not found.
Incorperated changes requested during review.