github-ci: use bundle.sh; better variable names
Use ./scripts/bundle.sh for bundling libhtp and suricata-update in
CI. This required a little bit of restructuring as the source code is
needed in the prep step to run bundle.sh, but overall I think its a
little cleaner.

Also gives better variable names for overrides in the pull request
message that should be more straight forward to use.