detect/dataset: delay set operation after signature full match
The set operation of dataset keyword was done even if signature
did not fully match.

This patch changes the behavior of the dataset keyword to do a
match and a post match for the set operation.
The postmatch retrieves the data and set it in the buffer.

Increases postmatch capability to do applayertxmatch,
and this get the data from a tx buffer.

Ticket: #5576