src: Use uniform type for tenant_id
SUMMARY:
The tenant_id attribute was defined as int in some files and as uint32_t in some other cases.
Switching to a uniform definition of uint32_t with this commit.

Fix formatting errors reported by clang-format

Fix data types for related variables

Fixed a comparison against tenant_id that had the type int. Converted it to uint32_t