decode-udp: Add exception for invalid length on padded packets
If the original packet is shorter than Ethernet's minimal transmission
unit (64B), the transmitting device pads the packet to the given size.
Padding is added to the IP layer but the UDP layer remains unaffected.
As a result, the UDP header length does not reach the end of the packet
as there is still padding after the UDP layer.

Redmine ticket: #5693